Problem 10: newly added attribute and enum were moved into Routing folder.
